More amd64 fixes to make it actually work
git-svn-id: http://svn.code.sf.net/p/navit/code/trunk/navit@112 ffa7fe5e-494d-0410-b361-a75ebd5db220

+++ src/block.h
@@ -2,9 +2,9 @@
 #include "transform.h"
 
 struct block {
-	unsigned long blocks;
-	unsigned long size;
-	unsigned long next;
+	u32 blocks;
+	u32 size;
+	u32 next;
 	struct coord c[2];
 	int count;
 };
